refactor(exception): customException을 상속하도록 구조 변경
