+++ src/main/java/woowacourse/exception/AuthException.java
@@ -2,15 +2,9 @@ package woowacourse.exception;
 
 import woowacourse.exception.dto.ErrorResponse;
 
-public class AuthException extends RuntimeException {
-    private final ErrorResponse errorResponse;
+public class AuthException extends CustomException {
 
     public AuthException(String message, ErrorResponse errorResponse) {
-        super(message);
-        this.errorResponse = errorResponse;
-    }
-
-    public ErrorResponse getErrorResponse() {
-        return errorResponse;
+        super(message, errorResponse);
     }
 }
