+++ src/main/java/woowacourse/exception/CustomException.java
@@ -1,0 +1,16 @@
+package woowacourse.exception;
+
+import woowacourse.exception.dto.ErrorResponse;
+
+public class CustomException extends RuntimeException {
+    private final ErrorResponse errorResponse;
+
+    public CustomException(String message, ErrorResponse errorResponse) {
+        super(message);
+        this.errorResponse = errorResponse;
+    }
+
+    public ErrorResponse getErrorResponse() {
+        return errorResponse;
+    }
+}
