+++ src/main/java/woowacourse/exception/InputFormatException.java
@@ -2,15 +2,9 @@ package woowacourse.exception;
 
 import woowacourse.exception.dto.ErrorResponse;
 
-public class InputFormatException extends RuntimeException {
-    private final ErrorResponse errorResponse;
+public class InputFormatException extends CustomException {
 
     public InputFormatException(String message, ErrorResponse errorResponse) {
-        super(message);
-        this.errorResponse = errorResponse;
-    }
-
-    public ErrorResponse getErrorResponse() {
-        return errorResponse;
+        super(message, errorResponse);
     }
 }
