+++ src/main/java/woowacourse/exception/JoinException.java
@@ -2,15 +2,9 @@ package woowacourse.exception;
 
 import woowacourse.exception.dto.ErrorResponse;
 
-public class JoinException extends RuntimeException {
-    private final ErrorResponse errorResponse;
+public class JoinException extends CustomException {
 
     public JoinException(String message, ErrorResponse errorResponse) {
-        super(message);
-        this.errorResponse = errorResponse;
-    }
-
-    public ErrorResponse getErrorResponse() {
-        return errorResponse;
+        super(message, errorResponse);
     }
 }
