+++ src/main/java/woowacourse/exception/LoginException.java
@@ -2,15 +2,9 @@ package woowacourse.exception;
 
 import woowacourse.exception.dto.ErrorResponse;
 
-public class LoginException extends RuntimeException {
-    private final ErrorResponse errorResponse;
+public class LoginException extends CustomException {
 
     public LoginException(String message, ErrorResponse errorResponse) {
-        super(message);
-        this.errorResponse = errorResponse;
-    }
-
-    public ErrorResponse getErrorResponse() {
-        return errorResponse;
+        super(message, errorResponse);
     }
 }
